He seguido actualizando el codigo en varios archivos .java

+++ cafeteria/src/lib/Productos.java
@@ -21,7 +21,7 @@ public class Productos {
 	public double getPrecio() {
 		return precio;
 	}
-	public void setPrecio(int precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
 	public String getTipo_Producto() {
